Edited database functions. They work so far now (tested).   * initialize database, deploy it if needed   * create a task and   * write it to database in connection with the main step while   * respecting the foreign-key constraints
In order to achieve the forth one, main_step attribute of FlowDB::Task now is_nullable => 1 so there is no circular, unresoluble dependency. It is now up to the wrappers for insert() and update (the latter not yet tested) to make sure the task is connected with the main step.