Shortened README.md and LIESMICH.md by a paragraph.
That paragraph still holds, but it does not quite fit in the README file, so I move it into the git commit history:

"I decided conciously to not make from FlowgencyTM yet another online site that is closed and central. Thus, maybe, I forgo earning good money with it. The question is if it were really good, considering the risk of only further strengthening people's dependency on machines."